Fix gas resistance reading
There is a report in the M5Stack community forum that ENVPro unit fails to report gas resistance. Apparently the returned number alwas is 0.0. So I looked into the source code and it looks like there might be a copy paste error here.

Please note: I do not have the ENVPro unit myself so I cannot test this. It's just a hunch that this might be the issue.